+++ components/Modal/ReadPost/hooks/useImageControl.ts
@@ -1,12 +1,8 @@
 import { TouchEventHandler, useState } from 'react';
 
-import useModalStore from '@stores/useModalStore';
-
-export default function useReadPost(imageList: string[] = []) {
-  const [isKebabClicked, setIsKebabClicked] = useState(false);
+export default function useImageControl(imageList: string[] = []) {
   const [currentImageIndex, setCurrentImageIndex] = useState(0);
   const [touchStartX, setTouchStartX] = useState(0);
-  const { toggleModal } = useModalStore();
 
   function handlePrevImage() {
     setCurrentImageIndex((prevIndex) => (prevIndex === 0 ? imageList.length - 1 : prevIndex - 1));
@@ -40,18 +36,11 @@ export default function useReadPost(imageList: string[] = []) {
     }
   };
 
-  function toggleKebab() {
-    setIsKebabClicked((prev) => !prev);
-  }
-
   return {
     currentImageIndex,
-    toggleModal,
     handleTouchStart,
     handleTouchEnd,
     handlePrevImage,
     handleNextImage,
-    isKebabClicked,
-    toggleKebab,
   };
 }
