+++ components/Modal/ReadPost/hooks/useKebabState.ts
@@ -1,0 +1,14 @@
+import { useState } from 'react';
+
+export default function useKebabState() {
+  const [isKebabClicked, setIsKebabClicked] = useState(false);
+
+  function toggleKebab() {
+    setIsKebabClicked((prev) => !prev);
+  }
+
+  return {
+    isKebabClicked,
+    toggleKebab,
+  };
+}
